refactor: Add useDialog hook
